Implement a MinStartTimeFilter
